Keep whitelist editor menu open when leaving hover area
