+++ app/src/marketplace/components/WhitelistEditor/index.jsx
@@ -1,0 +1,267 @@
+// @flow
+
+import React, { useCallback, useState } from 'react'
+import styled, { css } from 'styled-components'
+import { I18n } from 'react-redux-i18n'
+
+import Button from '$shared/components/Button'
+import UnstyledToggle from '$shared/components/Toggle'
+import UnstyledPopover from '$shared/components/Popover'
+import useCopy from '$shared/hooks/useCopy'
+import StatusIcon from '$shared/components/StatusIcon'
+import type { WhitelistItem } from '$mp/modules/contractProduct/types'
+import Notification from '$shared/utils/Notification'
+import { NotificationIcon } from '$shared/utils/constants'
+import type { Address } from '$shared/flowtype/web3-types'
+
+const MIN_ROWS = 5
+
+const Rows = styled.div`
+    max-height: ${({ theme }) => (((theme.rowCount) * 56) - 1)}px; /* +1 for header, -1 for bottom border */
+    overflow: auto;
+`
+
+const Popover = styled(UnstyledPopover)`
+    visibility: hidden;
+`
+
+const TableColumn = styled.div`
+    display: flex;
+    justify-content: center;
+    align-items: center;
+    padding: 0 24px;
+    color: rgba(82, 82, 82);
+
+    &:not(:last-child) {
+        border-right: 1px solid #ebebeb;
+    }
+`
+
+const TableRow = styled.div`
+    display: grid;
+    grid-template-columns: 1fr 90px 90px;
+    height: 56px;
+    background: #FDFDFD;
+
+    :hover {
+        background-color: #F8F8F8;
+
+        ${Popover} {
+            visibility: visible;
+        }
+    }
+
+    &:not(:last-child) {
+        border-bottom: 1px solid #ebebeb;
+    }
+
+    > *:first-child {
+        justify-content: unset;
+    }
+
+    ${({ theme }) => !!theme.disabled && css`
+        ${TableColumn} {
+            color: rgba(82, 82, 82, 0.5);
+        }
+
+        ${Popover} {
+            visibility: hidden;
+        }
+
+        :hover {
+            background-color: inherit;
+        }
+    `}
+
+    ${({ theme }) => !!theme.menuOpen && css`
+        ${Popover} {
+            visibility: visible;
+        }
+    `}
+`
+
+const TableHeaderRow = styled(TableRow)`
+    border-top-left-radius: 4px;
+    border-top-right-radius: 4px;
+
+    ${TableColumn} {
+        font-weight: 500;
+        letter-spacing: 0px;
+    }
+`
+
+const Controls = styled.div`
+    height: 72px;
+    padding: 0 16px 0 24px;
+    border-top: 1px solid #ebebeb;
+    display: grid;
+    grid-template-columns: auto 1fr auto;
+    grid-column-gap: 16px;
+    align-items: center;
+`
+
+const Toggle = styled(UnstyledToggle)`
+    display: flex;
+    align-items: center;
+
+    * {
+        margin-bottom: 0;
+    }
+`
+
+const Label = styled.label`
+    margin-bottom: 0;
+`
+
+type CommonProps = {
+    enabled: boolean,
+    onRemove: (Address) => Promise<void>,
+    actionsEnabled: boolean,
+}
+
+type EditorProps = CommonProps & {
+    items: Array<WhitelistItem>,
+    onEnableChanged: (boolean) => void,
+    onAdd: () => Promise<void>,
+}
+
+type RowProps = CommonProps & {
+    item: WhitelistItem,
+}
+
+const padWithEmptyRows = (rows: Array<WhitelistItem>) => {
+    if (rows.length < MIN_ROWS) {
+        const empties = new Array(MIN_ROWS - rows.length).fill(null)
+        return empties.map((item, index) => (
+            // eslint-disable-next-line react/no-array-index-key
+            <TableRow key={index}>
+                <TableColumn />
+                <TableColumn />
+                <TableColumn />
+            </TableRow>
+        ))
+    }
+
+    return null
+}
+
+const statusIconTheme = (disabled, status) => {
+    if (disabled) {
+        return StatusIcon.INACTIVE
+    } else if (status === 'added') {
+        return StatusIcon.PENDING
+    } else if (status === 'removed') {
+        return StatusIcon.REMOVED
+    }
+    return StatusIcon.OK
+}
+
+const TableRowItem = ({ item, enabled, actionsEnabled, onRemove }: RowProps) => {
+    const [menuOpen, setMenuOpen] = useState(false)
+    const disabled = !enabled || item.isPending
+
+    const { copy } = useCopy()
+
+    const onCopy = useCallback((address: string) => {
+        copy(address)
+
+        Notification.push({
+            title: I18n.t('general.copied'),
+            icon: NotificationIcon.CHECKMARK,
+        })
+    }, [copy])
+
+    return (
+        <TableRow
+            theme={{
+                disabled,
+                menuOpen,
+            }}
+        >
+            <TableColumn>
+                <span>{item.address}</span>
+            </TableColumn>
+            <TableColumn>
+                <StatusIcon
+                    status={statusIconTheme(disabled, item.status)}
+                    tooltip={!disabled && I18n.t(`editProductPage.whitelist.status.${item.status}`)}
+                />
+            </TableColumn>
+            <TableColumn>
+                <Popover
+                    title="Select"
+                    type="meatball"
+                    noCaret
+                    onMenuToggle={setMenuOpen}
+                >
+                    <Popover.Item onClick={() => onCopy(item.address)}>
+                        {I18n.t('editProductPage.whitelist.copy')}
+                    </Popover.Item>
+                    {item.status !== 'removed' && actionsEnabled && !disabled && (
+                        <Popover.Item onClick={() => onRemove(item.address)}>
+                            {I18n.t('editProductPage.whitelist.remove')}
+                        </Popover.Item>
+                    )}
+                </Popover>
+            </TableColumn>
+        </TableRow>
+    )
+}
+
+export const UnstyledWhitelistEditor = ({
+    enabled,
+    items,
+    onEnableChanged,
+    onAdd,
+    onRemove,
+    actionsEnabled,
+    ...props
+}: EditorProps) => (
+    <div {...props}>
+        <TableHeaderRow>
+            <TableColumn>{I18n.t('editProductPage.whitelist.header.address')}</TableColumn>
+            <TableColumn>{I18n.t('editProductPage.whitelist.header.status')}</TableColumn>
+            <TableColumn />
+        </TableHeaderRow>
+        <Rows theme={{
+            rowCount: MIN_ROWS,
+        }}
+        >
+            {items.map((item) => (
+                <TableRowItem
+                    key={item.address}
+                    item={item}
+                    enabled={enabled}
+                    actionsEnabled={actionsEnabled}
+                    onRemove={onRemove}
+                />
+            ))}
+            {padWithEmptyRows(items)}
+        </Rows>
+        <Controls>
+            <Label htmlFor="whitelist">{I18n.t('editProductPage.whitelist.enable')}</Label>
+            <Toggle
+                id="whitelist"
+                value={enabled}
+                onChange={onEnableChanged}
+            />
+            <Button
+                kind="secondary"
+                size="normal"
+                disabled={!enabled || !actionsEnabled}
+                onClick={() => onAdd()}
+            >
+                {I18n.t('editProductPage.whitelistEdit.add')}
+            </Button>
+        </Controls>
+    </div>
+)
+
+const WhitelistEditor = styled(UnstyledWhitelistEditor)`
+    background: #fdfdfd;
+    border-radius: 4px;
+    border: 1px solid #ebebeb;
+    font-size: 14px;
+`
+
+export default WhitelistEditor
